pyproject.toml
[tool.poetry]
name = "karaoke-gen"
version = "0.50.0"
description = "Generate karaoke videos with synchronized lyrics. Handles the entire process from downloading audio and lyrics to creating the final video with title screens."
authors = ["Andrew Beveridge <andrew@beveridge.uk>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "karaoke_prep" }]
homepage = "https://github.com/karaokenerds/karaoke-gen"
repository = "https://github.com/karaokenerds/karaoke-gen"
documentation = "https://github.com/karaokenerds/karaoke-gen/blob/main/README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = "<2.5"
requests = ">=2"
beautifulsoup4 = ">=4"
yt-dlp = "*"
lyricsgenius = ">=3"
fetch-lyrics-from-genius = ">=0.1"
pillow = ">=10.1"
pyinstaller = ">=6.3"
google-api-python-client = "*"
google-auth = "*"
google-auth-oauthlib = "*"
google-auth-httplib2 = "*"
thefuzz = ">=0.22"
numpy = "^1"
audio-separator = { version = ">=0.21.0", extras = ["cpu"] }
lyrics-converter = ">=0.2.1"
lyrics-transcriber = ">=0.34"
kbputils = "^0.0.16"
attrs = ">=24.2.0"
cattrs = ">=24.1.2"
toml = ">=0.10"
argparse = ">=1.4.0"
psutil = "^7.0.0"
pyperclip = "*"

[tool.poetry.group.dev.dependencies]
black = ">=23"
poetry = "*"

[tool.poetry.scripts]
karaoke-gen = 'karaoke_prep.cli:main'
karaoke-bulk = 'karaoke_prep.cli.bulk_cli:main'
# Legacy entry points for backward compatibility
karaoke-gen-legacy = 'karaoke_prep.utils.gen_cli:main'
karaoke-bulk-legacy = 'karaoke_prep.utils.bulk_cli:main'

[tool.black]
line-length = 140

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
